+++ exam/src/main/java/com/ustudy/exam/dao/StudentPaperDao.java
@@ -10,5 +10,5 @@ public interface StudentPaperDao {
 
 	void insertStudentPaper(StudentPaper paper);
 	
-	void deleteStudentPapers(@Param("csId")Integer csId, @Param("batchNum")Integer batchNum);
+	void deleteStudentPapers(@Param("csId")Long csId, @Param("batchNum")Integer batchNum);
 }
